add run labels to output and add record removal count

+++ cmd/dudu/clean.go
@@ -16,21 +16,31 @@ func init() {
 
 var cleanCmd = &cobra.Command{
 	Use:   "clean <path>",
-	Short: "Remove data from previous runs",
+	Short: "Remove records of previous runs",
 	Run: func(cmd *cobra.Command, args []string) {
+		count := 0
 		allFlag, _ := cmd.Flags().GetBool("all")
 		if allFlag == true {
 			files, _ := os.ReadDir(TempDir)
 			for _, file := range files {
 				os.Remove(TempDir + "/" + file.Name())
+				count++
 			}
+			fmt.Printf("Removed %v records\n", count)
 		} else {
-			filePaths, err := filepath.Glob(TempDir + "/dudu" + (strings.ReplaceAll(args[0], "/", ".")) + ".*")
-			if err != nil {
-				fmt.Println(err)
-			}
-			for _, file := range filePaths {
-				os.Remove(file)
+			if len(args) < 1 {
+				fmt.Println("Please specify a run path to clean")
+				return
+			} else {
+				filePaths, err := filepath.Glob(TempDir + "/dudu" + (strings.ReplaceAll(args[0], "/", ".")) + ".*")
+				if err != nil {
+					fmt.Println(err)
+				}
+				for _, file := range filePaths {
+					os.Remove(file)
+					count++
+				}
+				fmt.Printf("Removed %v records\n", count)
 			}
 		}
 	},
